Implemented multithreaded version for improved performance.

+++ Utilities/io.cpp
@@ -39,7 +39,7 @@ void readImageFileToVector(std::string fileName, std::vector<uint8_t> *data) {
   @param WIDTH width in bytes of the output image
   @param HEIGHT height in bytes of the output image
  */
-void writeDataVectorToFile( std::string fileName, std::vector<uint8_t> *data, const int WIDTH, const int HEIGHT ) {
+void writeDataVectorToFile( std::string fileName, std::vector<uint8_t> *data, const size_t WIDTH, const size_t HEIGHT ) {
 	std::fstream image_file_stream (fileName, std::ios::out | std::ios::binary);
     char* binValArray = new char [WIDTH * HEIGHT];
 	std::copy( data->begin(), data->end(), binValArray );
